requirements.txt
watchdog
pdfplumber
pytesseract
pdf2image
Pillow
mysql-connector-python
python-dotenv
fastapi
uvicorn[standard]
openai>=1.40.0
